+++ apps/xdhdq/frontend/XDHTML/Frame.h
@@ -30,15 +30,15 @@ namespace frame {
 		BASE_ACR( Template );
 	};
 
-	void SetLayout( const char * Id, core::rSession & Session );
+	void SetLayout( const char * Id, core::rSession_ & Session );
 
 	void SetCasting(
 		const char *Id,
-		core::rSession &Session );
+		core::rSession_ &Session );
 
 	void Display(
 		const char *Id,
-		core::rSession &Session );
+		core::rSession_ &Session );
 }
 
 #endif