+++ apps/xdhdq/frontend/XDHTML/login.h
@@ -35,11 +35,11 @@ namespace login {
 		BASE_ACR( Dismiss );
 	}
 
-	void SetLayout( core::rSession & Session );
+	void SetLayout( core::rSession_ & Session );
 
-	void SetCasting( core::rSession & Session );
+	void SetCasting( core::rSession_ & Session );
 
-	void Display( core::rSession &Session );
+	void Display( core::rSession_ &Session );
 }
 
 #endif