+++ stable/xdhujr.cpp
@@ -57,8 +57,8 @@ entry___ xdhujr::script::value::Getter( "Getter", ValueScripts_ );
 
 static entry___ FetcherScripts_("Fetcher", xdhujr::Scripts );
 entry___ xdhujr::script::fetcher::Events( "Events", FetcherScripts_ );
+entry___ xdhujr::script::fetcher::Widgets( "Widgets", FetcherScripts_ );
 entry___ xdhujr::script::fetcher::Casts( "Casts", FetcherScripts_ );
-entry___ xdhujr::script::fetcher::ContentsAndWidgets("ContentsAndWidgets", FetcherScripts_ );
 
 static entry___ SetterScripts_( "Setter", xdhujr::Scripts );
 entry___ xdhujr::script::setter::EventHandlers( "EventHandlers", SetterScripts_ );